adds a few tests for the utils_dict
